requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
arxiv==2.1.0
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
Brotli==1.1.0
bs4==0.0.2
build==1.2.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.1
cssselect==1.2.0
cssselect2==0.7.0
curl_cffi==0.6.3
cycler==0.12.1
dataclasses-json==0.6.4
decorator==5.1.1
Deprecated==1.2.14
diskcache==5.6.3
distro==1.9.0
docker==7.0.0
docopt==0.6.2
duckduckgo_search==5.3.0
executing==2.0.1
fastapi==0.110.2
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.13.4
FLAML==2.1.2
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
google-ai-generativelanguage==0.6.2
google-api-core==2.19.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.2
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.2
grpcio-status==1.62.2
h11==0.14.0
html5lib==1.1
htmldocx==0.0.6
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.25.2
huggingface-hub==0.22.2
humanfriendly==10.0
idna==3.7
importlib-metadata==7.0.0
importlib_resources==6.4.0
iniconfig==2.0.0
ipython==8.23.0
jedi==0.19.1
jieba3k==0.35.1
Jinja2==3.1.3
joblib==1.4.0
jsonpatch==1.33
jsonpointer==2.4
kiwisolver==1.4.5
kubernetes==29.0.0
langchain==0.1.16
langchain-community==0.0.34
langchain-core==0.1.48
langchain-google-genai==1.0.3
langchain-openai==0.1.5
langchain-text-splitters==0.0.1
langsmith==0.1.50
lxml==5.2.1
lxml_html_clean==0.1.1
Markdown==3.6
markdown-it-py==3.0.0
markdown2==2.4.13
markdownify==0.12.1
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.4
matplotlib-inline==0.1.7
md2pdf==1.0.1
mdurl==0.1.2
mistune==3.0.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.3
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.17.3
openai==1.20.0
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
orjson==3.10.1
overrides==7.7.0
packaging==23.2
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
playwright==1.43.0
pluggy==1.5.0
posthog==3.5.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pulsar-client==3.5.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyautogen==0.2.27
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
pydyf==0.10.0
pyee==11.1.0
Pygments==2.17.2
PyMuPDF==1.24.2
PyMuPDFb==1.24.1
pyparsing==3.1.2
pypdf==4.2.0
pyphen==0.15.0
PyPika==0.48.9
pyproject_hooks==1.0.0
pytest==8.1.1
pytest-asyncio==0.23.6
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.9
python-ulid==2.4.0.post0
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
requests-file==2.0.0
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
sentence-transformers==2.7.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
stack-data==0.6.3
starlette==0.37.2
sympy==1.12
tavily-python==0.3.3
tenacity==8.2.3
termcolor==2.4.0
threadpoolctl==3.4.0
tiktoken==0.6.0
tinycss2==1.3.0
tinysegmenter==0.3
tldextract==5.1.2
tokenizers==0.19.1
torch==2.2.2
tqdm==4.66.2
traitlets==5.14.3
transformers==4.40.0
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.13
weasyprint==62.0
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
youtube-transcript-api==0.6.2
zipp==3.18.1
zopfli==0.2.3
